work in progress on conntrack monitor (Comment: #12984)
git-svn-id: https://ut.svn.beanstalkapp.com/ngfw/work/src@43477 8ebf8b4e-06f3-0310-961f-89f74618be54

+++ jnetcap/impl/com/untangle/jnetcap/Netcap.java
@@ -37,7 +37,7 @@ public final class Netcap
     /**
      * Return the next available session ID
      */
-    public static native int nextSessionId();
+    public static native long nextSessionId();
 
     /** 
      * Initialize the JNetcap and Netcap library with the same debugging level for
